requirements.txt
Django==4.2.10
djongo==1.3.6
pymongo==4.6.1
python-dateutil==2.8.2
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
sqlparse==0.2.4 